refactor(pipeline/pii): split Amount + HTML redactors into per-category modules
Phase 6 commit 4/7. Extracts redactAmount + redactHtml (incl. replaceValueAttr, HTML_TEXT_PATTERNS, HTML_VALUE_ATTR_RE, coerceToNumber) into dedicated PiiRedactor/Amount.ts and PiiRedactor/Html.ts modules. PiiRedactor.ts re-exports the moved symbols so the public API is byte-identical; dispatchStrategy now imports redactAmount from the new module.
